cleanup, breakout toolbaractions.go
